dev page2 refactor to page0_template
