+++ src/main/java/graphics/devMode/page0_template/ReadMe.md
@@ -1,0 +1,9 @@
+# Page 0
+
+## Abstract
+
+A template for new dev pages.
+
+## Instructions
+
+Copy-paste this and improve.
