--- src/main/java/graphics/devMode/page2/ReadMe.md
@@ -1,11 +1,0 @@
-# Page 2
-
-## Abstract
-
-_A placeholder._
-
-_Coming soon..._
-
-## Instructions
-
-_Coming soon..._
