--- src/main/java/graphics/devMode/selectionPage/buttons/Button_Page2.java
@@ -1,19 +1,0 @@
-package graphics.devMode.selectionPage.buttons;
-
-import java.awt.event.ActionListener;
-
-import org.jetbrains.annotations.Nullable;
-
-import graphicsEngine.parts.SimpleJButton;
-
-/**
- * TODO: add javadocs
- */
-public class Button_Page2 extends SimpleJButton {
-    protected static final String ACTION_COMMAND = "page_2";
-
-    public Button_Page2(@Nullable ActionListener actionListener) {
-        super("Page2", ACTION_COMMAND, actionListener);
-        setFocusable(false);
-    }
-}
